Move some components out of utils
